chore: move service config dao to api module
