+++ persistence-api/src/commonMain/kotlin/com/wire/kalium/persistence/model/ServerConfigWithUserIdEntity.kt
